ImageDownloadError вынесен в отдельный класс
